+++ CHANGELOG.md
@@ -3,12 +3,16 @@
 
 ## Unreleased
 
+### Added
+
+- MockClientStrategy class.
+
 ## 1.1.1 - 2016-11-27
 
 ### Changed
 
 - Made exception messages clearer. `StrategyUnavailableException` is no longer the previous exception to `DiscoveryFailedException`.
-- `CommonClassesStrategy` is using `self` instead of `static`. Using `static` makes no sense when `CommonClassesStrategy` is final. 
+- `CommonClassesStrategy` is using `self` instead of `static`. Using `static` makes no sense when `CommonClassesStrategy` is final.
 
 ## 1.1.0 - 2016-10-20